requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
google-generativeai
py-trello
schedule
requests-oauthlib